backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.6,<5.0
google-auth>=2.25,<3.0
requests>=2.31,<3.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
razorpay>=1.4,<2.0
django-storages>=1.14,<2.0
boto3>=1.34,<2.0
Pillow>=10.0,<11.0